move indexFromCsv helper method to GroupMappingServiceTest
